requirements.txt
numpy==1.16.2
cython==0.28.5
tensorboard==1.8.0
tensorboardX==1.6
imageio==2.6.1
scikit-image==0.14.0
pydensecrf==1.0rc2
torch==0.4.1
torchvision==0.2.1
scipy==1.1.0
opencv-python==3.4.2.16
opencv-contrib-python == 3.4.2.16
pandas==0.23.4
Pillow==5.2.0
mxnet
